Removed guard
This may be a point of contention, but here's the thing ... we
want compatibility for those still using ruby-1.9, so we need to
test on it. Unfortunately, guard requires listen, which requires
ruby_dep, which is apparently ruby-2.X-only.

So, removing the guard test dependency, becuase it's not hard
to just run `bundle exec rake` every now and again.