Fixes regression from ENG-965
In #68 , the behavior of `ey-core ssh` was inadvertently changed
for the non-command (open a shell) case, but this behavior was
not documented.

This commit doesn't revert the change from #68, but does restore
the default behavior such that if no command is passed and no
role is passed, the app_master/solo instance in the environment
is explicitly added to the list of servers to use for the
connection.