Fix account and environment filters
The previous iteration returned an error when specifying an account. Furthermore, when a user has access to multiple accounts with multiple environments of the same name, the environments toggle would return only return the environment for the main account—and since the account couldn’t be set, this was problematic. Also, the environments short option has been fixed (“-e” has been changed to “e”).